CantCanciones arreglado: bst para no repetir canciones en la cuenta

+++ src/uy/edu/um/main/Entities/ReportsInterface.java
@@ -3,6 +3,7 @@ package uy.edu.um.main.Entities;
 import uy.edu.um.adt.hash.KeyNullException;
 import uy.edu.um.main.Exceptions.DatosInvalidosException;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 
 public interface ReportsInterface {
@@ -14,5 +15,5 @@ public interface ReportsInterface {
 
     public int cantArtista(String a, String d, String p) throws DatosInvalidosException, KeyNullException;
 
-    public int cantCanciones(Double ti, Double tf, String di, String df) throws DatosInvalidosException, KeyNullException;
+    public int cantCanciones(BigDecimal ti, BigDecimal tf, String di, String df) throws DatosInvalidosException, KeyNullException;
 }
